Fall back to first option of NOVOPlasty if not circular
If `NOVOPlasty` is unable to find a complete circular chloroplast the file `Circularized_assembly_1_NOVOPlasty.fasta` is not created at all. According to the [documentation](https://github.com/ndierckx/NOVOPlasty#4-output-files) different options are returned with names like `Option_nr_projectname.txt`. So we can just copy the first option as final result. This behavior is not perfect as `NOVOPlasty` does not promote this as the best option or an assumed final chloroplast. Still for our benchmarking we can consider this as the best guess rather than assuming a complete failure on the dataset.

Fix #12